backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pytest==8.0.2
flake8==7.0.0
black==24.2.0
isort==5.13.2
lxml==5.1.0
psycopg2-binary==2.9.9 